feat: upload files and load file by folder

+++ src/main/java/com/myfolder/myfolder/domain/entities/FileEntity.java
@@ -2,6 +2,6 @@ package com.myfolder.myfolder.domain.entities;
 
 import java.util.UUID;
 
-public record FileEntity(UUID id, String filename, Double fileLength, String type, Boolean isSafe, Boolean isDeleted,
+public record FileEntity(UUID id, String filename, Long fileLength, String type, Boolean isSafe, Boolean isDeleted,
                          UUID folderId, String path) {
 }
